+++ packages/sdk-redux/src/reduxSlices/transactionTrackerSlice/transactionTrackerAdapter.ts
@@ -4,7 +4,6 @@ import {TrackedTransaction} from './trackedTransaction';
 import {TransactionTrackerReducer, transactionTrackerSlicePrefix} from './transactionTrackerSlice';
 
 export const transactionTrackerAdapter = createEntityAdapter<TrackedTransaction>({
-    selectId: (transaction) => transaction.hash,
     sortComparer: (a) => a.timestampMs,
 });